[FIX] 地址更新

+++ src/main/java/org/srm/mall/other/api/dto/WatsonsAddressDTO.java
@@ -45,7 +45,7 @@ public class WatsonsAddressDTO {
     /**
      * 商城对应的地区等级
      */
-    private Integer mallRegionLevel;
+    private Long mallRegionLevel;
 
     /**
      * 百度地区编码
